worker: drain via USR1/USR2
right now it's way too easy to end up with a drain failure; deleting the
deployment, for example, will cause the 'land-worker/retire-worker'
command to fail, if the TSA goes away before the draining begins.

similarly, if the land or retire command runs but then the TSA goes away
before draining completes, the 'worker' process will stick around
trying to reconnect, not realizing that it's been drained.

the drain script also just isn't given enough information to know
whether it needs to land or retire. we'll have to open an issue with
BOSH or something to figure that out. we need to know if the VM is
physically going away, and there's no unambiguous way to detect that
(`bosh stop` results in job_shutdown as well).

yet another issue: the 'worker' process knows to exit when it
disappears, which isn't exactly the desired behavior if the worker is
marked 'ephemeral' and stalled (which results in its removal).

to fix this, we can send USR1 (land) or USR2 (retire) to the process
itself, which will allow it to know that it's being shut down. it can
initiate the land/retire itself, and at least know that if the TSAs are
all gone that it should probably shut down instead of trying to
reconnect forever.