adding some unit tests for envgen's ruby helpers
this is pretty shoddy and requires temporarily pasting the ruby code
under test into this file to test it.

but since:

1) neither this release nor envgen has a test suite, and
2) I don't expect this code to change very often,

I don't think it's worth putting in more work to do this in a more
complicated way that allows the ruby helpers to be in a separate file
that's explicitly under test.

Signed-off-by: Mike Dalessio <mike.dalessio@gmail.com>