remove guardian flag defaults
CONCOURSE_GARDEN_ALLOW_HOST_ACCESS will have a value of "false"
even without a default value being provided as it is represented
as a bool in the Guardian codebase.

The recommendation to have CONCOURSE_GARDEN_MAX_CONTAINERS at 250
was based off CloudFoundry's Diego which also uses Guardian.
Concourse deployments have been operated at a much higher limit
and do not need to be held to this limit.

The binary will set CONCOURSE_GARDEN_NETWORK_POOL to a default value of
10.80.0.0/16 as of concourse/concourse#6031.

concourse/concourse#5985

Signed-off-by: Muntasir Chowdhury <mchowdhury@pivotal.io>