add CONCOURSE_CONTAINERD_INIT_BIN
The CONCOURSE_CONTAINERD_INIT_BIN needs to point to
/var/vcap/packages/concourse/bin/init specifically on BOSH when running
containerd.

This should only be set when using containerd as the runtime. Currently,
that means checking if the runtime property is set to "containerd"
explicitly. In the future when the concourse binary will default to
containerd when no runtime property is set we also need to check if
runtime is nil.

concourse/concourse#5984

Signed-off-by: Muntasir Chowdhury <mchowdhury@pivotal.io>